fix: Add my user to the video group
Some time ago I cleaned up the extraGroups for my user and some stuff
stopped working. Of course I did not realize that it was because of
the group change, otherwise I would have fixed it directly.

The program that I missed and which stopped working was `light`. It
has a NixOS module which puts a setuid wrapper to the binary that
marks it as executable _for users in the video group_.