fix(users/rofi): address rofi-pass bug
There is a bug in the combination of rofi-pass 1.5.3 and rofi 1.7.5 where an option that was used in rofi-pass has been deprecated, thus
causing rofi to show a confused prompt asking us what we're trying to do.

The suggested fix in the [issue] avoids this code path by setting the help_color to some non-empty value.

[issue]: https://github.com/carnager/rofi-pass/issues/226